feat:DTOSS-7656 Sonar cloud issue fix
