+++ application/CohortManager/src/Functions/CaasIntegration/RetrieveMeshFile/HealthCheckFunction.cs
@@ -5,6 +5,7 @@ using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
 using System.Net;
 using System.Threading.Tasks;
+using HealthChecks.Extensions;
 
 public class HealthCheckFunction
 {
@@ -18,15 +19,6 @@ public class HealthCheckFunction
     [Function("health")]
     public async Task<HttpResponseData> Run([HttpTrigger(AuthorizationLevel.Function, "get")] HttpRequestData req)
     {
-        var healthReport = await _healthCheckService.CheckHealthAsync();
-
-        var response = req.CreateResponse(healthReport.Status == HealthStatus.Healthy ? HttpStatusCode.OK : HttpStatusCode.ServiceUnavailable);
-        await response.WriteAsJsonAsync(new
-        {
-            status = healthReport.Status.ToString(),
-            details = healthReport.Entries
-        });
-
-        return response;
+        return await HealthCheckServiceExtensions.CreateHealthCheckResponseAsync(req, _healthCheckService);
     }
 }
